Moved aws-utilities to their own tree.  Updated tests, removed any SDK1 leftovers in dynamodb, confirmed lambda debug logs in docker-compose with localstack.
