+++ development-test/jar-lambda-poc/src/test/docker/docker-compose.yml
@@ -4,10 +4,10 @@ services:
     ports:
       - "4566:4566"
     environment:
-      - DEBUG=${DEBUG- }
+      - DEBUG=1
       - SERVICES=s3,lambda,events,logs
-#      - LAMBDA_DOCKER_FLAGS=-p 127.0.0.1:5050:5050 -e LOCALSTACK_INIT_LOG_LEVEL=debug
-      - LAMBDA_DOCKER_FLAGS=-e LOCALSTACK_INIT_LOG_LEVEL=debug
+#      - LAMBDA_DOCKER_FLAGS=-p 127.0.0.1:5050:5050 -e LS_LOG=debug
+      - LAMBDA_DOCKER_FLAGS=-e LS_LOG=debug
       - LAMBDA_RUNTIME_ENVIRONMENT_TIMEOUT=300
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
