+++ jar-development/pom.xml
@@ -27,6 +27,13 @@
         <relativePath>../java-development</relativePath>
     </parent>
 
+    <properties>
+
+        <!-- Excluded from unit test coverage by default as Spring configurations -->
+        <coverage.exclude.build.pattern.1>**.*Config</coverage.exclude.build.pattern.1>
+        <coverage.exclude.build.pattern.2>**.*Configuration</coverage.exclude.build.pattern.2>
+    </properties>
+
     <build>
         <!-- We define common PROCESS here.  The CONFIGURATIONS are in the parent pom -->
         <plugins>
