+++ jar-lambda-development/pom.xml
@@ -59,14 +59,8 @@
             <artifactId>spring-cloud-function-adapter-aws</artifactId>
         </dependency>
         <dependency>
-            <!-- Localstack support -->
-            <groupId>com.limemojito.oss.standards</groupId>
-            <artifactId>aws-utilities</artifactId>
-            <version>14.2.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.limemojito.oss.standards</groupId>
-            <artifactId>aws-lambda</artifactId>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>lambda-utilities</artifactId>
             <version>14.2.0-SNAPSHOT</version>
         </dependency>
