+++ spring-boot-development/pom.xml
@@ -43,11 +43,6 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.limemojito.oss.standards</groupId>
-            <artifactId>aws-utilities</artifactId>
-            <version>14.2.0-SNAPSHOT</version>
-        </dependency>
     </dependencies>
 
     <build>
