+++ utilities/aws-utilities/lambda-utilities/src/main/support/assembly/native.xml
