+++ utilities/aws-utilities/lambda-utilities/src/main/support/shell/bootstrap
