+++ utilities/aws-utilities/sqs-utilities/src/test/java/com/limemojito/aws/sqs/pump/TestMessage.java
