+++ utilities/aws-utilities/ssm-utilities/pom.xml
@@ -20,8 +20,10 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.limemojito.oss.standards.lock</groupId>
-    <artifactId>lock-dynamodb</artifactId>
+
+    <groupId>com.limemojito.oss.standards.aws</groupId>
+    <artifactId>ssm-utilities</artifactId>
+    <version>14.2.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <parent>
@@ -31,24 +33,14 @@
         <relativePath>../../../jar-development/pom.xml</relativePath>
     </parent>
 
-    <properties>
-        <coverage.line.ratio>0.71</coverage.line.ratio>
-    </properties>
-
     <dependencies>
         <dependency>
-            <groupId>com.limemojito.oss.standards.lock</groupId>
-            <artifactId>lock-api</artifactId>
-            <version>14.2.0-SNAPSHOT</version>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.limemojito.oss.standards.aws</groupId>
-            <artifactId>dynamodb-utilities</artifactId>
-            <version>14.2.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>dynamodb-lock-client</artifactId>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>ssm</artifactId>
             <exclusions>
                 <exclusion>
                     <groupId>commons-logging</groupId>
