+++ utilities/test-utilities/pom.xml
@@ -43,120 +43,28 @@
 
     <dependencies>
         <dependency>
-            <!-- Note this means that the aws-utilities framework can NOT depend on the test-framework -->
-            <groupId>com.limemojito.oss.standards</groupId>
-            <artifactId>aws-utilities</artifactId>
-            <version>${project.version}</version>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>io.awspring.cloud</groupId>
-            <artifactId>spring-cloud-aws-sqs</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>io.awspring.cloud</groupId>
-            <artifactId>spring-cloud-aws-sns</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>software.amazon.awssdk</groupId>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
-            <artifactId>s3</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>software.amazon.awssdk</groupId>
-            <artifactId>lambda</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>commons-logging</groupId>
-                    <artifactId>commons-logging</artifactId>
-                </exclusion>
-            </exclusions>
-            <optional>true</optional>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-json</artifactId>
-            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-validation</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>net.javacrumbs.json-unit</groupId>
-            <artifactId>json-unit-assertj</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-lambda-java-events</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-lambda-java-serialization</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>software.amazon.awssdk</groupId>
-            <artifactId>dynamodb-enhanced</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <!-- Prefer AssertJ over Hamcrest -->
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.hamcrest</groupId>
-            <artifactId>hamcrest-core</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.hamcrest</groupId>
-            <artifactId>hamcrest-library</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-webflux</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.freemarker</groupId>
             <artifactId>freemarker</artifactId>
-            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>net.javacrumbs.json-unit</groupId>
+            <artifactId>json-unit-assertj</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
-            <artifactId>spring-context</artifactId>
-            <optional>true</optional>
+            <artifactId>spring-webflux</artifactId>
         </dependency>
         <dependency>
             <groupId>commons-io</groupId>
@@ -173,10 +81,41 @@
             <artifactId>awaitility</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>lambda-utilities</artifactId>
+            <version>14.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>com.limemojito.oss.standards</groupId>
             <artifactId>json-utilities</artifactId>
             <version>14.2.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>dynamodb-utilities</artifactId>
+            <version>14.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>s3-utilities</artifactId>
+            <version>14.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>sqs-utilities</artifactId>
+            <version>14.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.limemojito.oss.standards.aws</groupId>
+            <artifactId>sns-utilities</artifactId>
+            <version>14.2.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <!-- Resist using this on new code -->
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+        </dependency>
+
 
         <!-- TEST DEPENDENCIES -->
         <dependency>
