+++ utilities/test-utilities/src/test/docker/docker-compose.yml
@@ -4,9 +4,9 @@ services:
     ports:
       - "4566:4566"
     environment:
-      - DEBUG=${DEBUG- }
-      #      - LAMBDA_DOCKER_FLAGS=-p 127.0.0.1:5050:5050 -e LOCALSTACK_INIT_LOG_LEVEL=debug
-      - LAMBDA_DOCKER_FLAGS=-e LOCALSTACK_INIT_LOG_LEVEL=debug
+      - DEBUG=1
+      #      - LAMBDA_DOCKER_FLAGS=-p 127.0.0.1:5050:5050 -e LS_LOG=debug
+      - LAMBDA_DOCKER_FLAGS=-e LS_LOG=debug
       - LAMBDA_RUNTIME_ENVIRONMENT_TIMEOUT=300
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
